requirements.txt
# Core dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
colorama>=0.4.6
requests>=2.28.0

# NFC dependencies
nfcpy>=1.0.4
pyusb>=1.2.1
pyscard>=2.0.7  # Alternative for macOS

# Google API dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# GUI dependencies (for cross-platform UI)
customtkinter>=5.2.0
Pillow>=10.0.0

# Additional utilities
python-dotenv>=1.0.0